refactor: extract getApiHeaders/getApiUrl into shared test helper
